Split fasta/fastq from view into its own source file.
It's been a bugbear of mine for ages!  Filename matches the prefix
used by most other commands, although not sam_view (another
annoyance).
