+++ sam_utils.c
@@ -23,6 +23,7 @@ FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 DEALINGS IN THE SOFTWARE.  */
 
 #include <config.h>
+#include <stdlib.h>
 
 #include <stdio.h>
 #include <stdarg.h>
@@ -57,4 +58,16 @@ void print_error_errno(const char *subcommand, const char *format, ...)
     va_start(args, format);
     vprint_error_core(subcommand, format, args, err? strerror(err) : NULL);
     va_end(args);
+}
+
+void check_sam_close(const char *subcmd, samFile *fp, const char *fname, const char *null_fname, int *retp)
+{
+    int r = sam_close(fp);
+    if (r >= 0) return;
+
+    // TODO Need error infrastructure so we can print a message instead of r
+    if (fname) print_error(subcmd, "error closing \"%s\": %d", fname, r);
+    else print_error(subcmd, "error closing %s: %d", null_fname, r);
+
+    *retp = EXIT_FAILURE;
 }
